+++ examples/google-quickdraw-cover-aggregator.js
@@ -3,7 +3,8 @@ import workerFn from './google-quickdraw-cover-aggregator-worker';
 
 const createGoogleQuickDrawCoverAggregator = ({
   size = 64,
-  lineWidth = 2
+  lineWidth = 2,
+  log = false
 } = {}) => items =>
   new Promise((resolve, reject) => {
     const worker = createWorker(workerFn);
@@ -17,7 +18,8 @@ const createGoogleQuickDrawCoverAggregator = ({
     worker.postMessage({
       lineWidth,
       items,
-      size
+      size,
+      log
     });
   });